Added send notifications for parents,rearranged bookings, notifications and parent home page code
